move PanelHeaderComponents
