+++ client/src/PanelHeaderComponents/Carousel.css
