+++ client/src/PanelHeaderComponents/Carousel.jsx
@@ -19,7 +19,7 @@ import "./Carousel.css";
 import { Box, Flex, IconButton, Text } from "@chakra-ui/react";
 import ContentCard from "./ContentCard";
 import { ChevronLeftIcon, ChevronRightIcon } from "@chakra-ui/icons";
-import { createFullName } from "../../_utils/names";
+import { createFullName } from "../_utils/names";
 
 SwiperCore.use([Keyboard, Mousewheel]);
