+++ client/src/Tools/_framework/Paths/Activities.jsx
@@ -30,7 +30,7 @@ import {
 import styled from "styled-components";
 
 import { RiEmotionSadLine } from "react-icons/ri";
-import ContentCard from "../../../_reactComponents/PanelHeaderComponents/ContentCard";
+import ContentCard from "../../../PanelHeaderComponents/ContentCard";
 import { GeneralActivityControls } from "./ActivityEditor";
 import axios from "axios";
